Removed weather logic adapter.
The weather logic adapter has a new home:
https://github.com/gunthercox/chatterbot-weather

It was removed from the main chatterbot project
to keep the number of requirements low. On top
of that, the weather logic adapter was also
completely untested, undocumented, and contained
multiple TODOs. In order to keep the quality of
chatterbot's software as high as possible this
adapter had to be pruned.

In the future it will be possible to download
the new chatterbot-weather module from PyPi.