Disconnect storage from adapter class
This makes a change to seperate the storage adapter from the base
Adapter class. This is being done because the methods on the base
Adapter class are not needed on the storage adapter.

The main change is that is no longer possible for the storage
adapter to access the ChatBot instance. This should not cause
any probems because there are no "known" use cases where this
is needed.

This also allows a special case to be removed in the class
validation utility so that a check to the "Adapter" class
is never made. This allows the class validation utility to
be more generic.